review: Move review records out of terminal rendering
File review defines the records for changes, pages, and the complete review
model inside the terminal-output package.

Those records only hold data and are also consumed by selection,
pagination, and saved review state. Locating them beside printing code
makes selection and saved-state code depend on terminal presentation.

This commit begins separating review data from output by moving the records
into data.file_review and updating every current import.

Later commits will move displayed-row mapping and selectable-row
calculations before saved review selections leave the output package.
