+++ src/git_stage_batch/output/file_review.py
@@ -8,7 +8,7 @@ from .colors import Colors
 from .file_review_action_selections import shown_line_action_selections
 from .file_review_display_ids import display_ids_for_rows
 from .file_review_footer import print_file_review_footer
-from .file_review_model import FileReviewModel
+from ..data.file_review.model import FileReviewModel
 from .file_review_rows import (
     maximum_display_id_digit_count,
     print_file_review_rows,
