+++ src/git_stage_batch/output/file_review_pagination.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from ..core.models import LineEntry
 from . import file_review_layout
-from .file_review_model import FileReviewPage, ReviewChange, ReviewChangeFragment
+from ..data.file_review.model import FileReviewPage, ReviewChange, ReviewChangeFragment
 
 
 def paginate_file_review_changes(
